Add withRecMap, for maps whose values can contain keys.
withRecMap is intended for use when the values in the map reference
keys that are supposed to also be in the map. For example, a directed
graph represented by Map Int [Int] can be used as a
Map ph [Int] [Key ph Int] within withRecMap, ensuring that all edges
have a target that exists in the map.